+++ main.py
@@ -5,7 +5,7 @@ Docstring for main
 import datetime
 from src import CONFIGURATION_JSON
 from src.sign_in import sign_in_interface
-from src.__etc__.activation import activation
+from src.__etc__._activation import activation
 
 _product_keys: list[str] = [
     "2030-GITH-UBGC-AKKI-DIST-FIRS-TPRJ-INDI-AUSA-2026",
