+++ src/__etc__/__init__.py
@@ -1,2 +1,2 @@
 from .. import *
-from .activation import activation
+from ._activation import activation
